Trying to get a clean compile on windows
A few of the internally forked crates had `deny(warnings)` configured
which as this code has been more-or-less unmaintained for ~2 years
that now was breaking builds.

An effort has been made to clean up these warnings, remove depreciated
language features, and re-format the sources.

Still a lot of work todo but this will help in the future.